resin-image: Adapt our image for the Jetson TX2
Implement python function to connect KERNEL_DEVICETREE variable to the deployed file,
this is very similiar to what kernel.bbclass does.
At present the BSP does not provide a bootloader for this board, we also do not have
a straight forward way to flash it if we had a bootloader.
The size is set in such a way that the total amount of space, that resin occupies (without the data partition)
, will equal 1GiB.

Signed-off-by: Theodor Gherzan <theodor@resin.io>