Install n510-tx2 dtb only for this machine
The device tree for the n510-tx2 machine seems to be loaded
by the first stage bootloader program(before u-boot) on the
jetson-tx2 instead of the jetson-tx2 device tree.
This patch installs the n510-tx2 device tree only when the machine
is built.

Changelog-entry: Install n510-tx2 dtb only for this machine
Signed-off-by: Sebastian Panceac <sebastian@balena.io>