bt-scripts: Fix bluetooth not working on TX2
These files are brought in from the Ubuntu demo image
provided by NVidia but are modified for our usecases.

We don't execute the bluetooth scripts from the udev
rule, instead we invoke them from the nvbtwifi* service
files. Also it is important to bring device up before
BlueZ gets the chance to interfere with it to avoid
intermittent failures.

Changelog-entry: Fix bluetooth not working on TX2
Signed-off-by: Alexandru Costache <alexandru@balena.io>