resin-image-initramfs: Install tegra-xusb firmware
Issuing lsusb will trigger loading of the tegra
xusb firmware, and the first call will fail. This because
when the driver was loaded it failed to load firmware from
the initramfs as it wasn't present, and after a couple
retries the process was aborted.

Install firmware in initramfs so that it can
be loaded when the driver loads and avoid
usb related issues.

Backport-to: next
Changelog-entry: resin-image-initramfs: Install tegra-xusb firmware
Signed-off-by: Alexandru Costache <alexandru@balena.io>