layers/meta-tegra: Switch BSP back to L4T 28.1
We do the switch to this version because we have been maintaining a
combination of newer L4t and older kernel which made some subtle bugs
pop up. So we switch all layers and components back to 28.1 which is
known to work for the customer.

Changelog-entry: Switch back to L4T version 28.1
Signed-off-by: Florin Sarbu <florin@balena.io>