linux-tegra: Remove unused patches
The patch 0001-bcmdhd-Fix-wifi-disconnect-problems.patch is not suitable
anymore for the L4T 28.1 kernel version so we remove the patch completely.

The patch 0001-Add-gasket-module-1.1.3.patch is for a machine we do not
have anymore so we remove this patch also.

The realsense patches are not used on the skx2 or the blackboard tx2 so
we remove them.

The 0001-Backport-qmi_wwan-from-kernel-4.14-to-4.4.patch was not applied
to the skx2 kernel before anyway and it would need to be redone if we
wanted so have it in so let's remove it.

Signed-off-by: Florin Sarbu <florin@balena.io>